Remove action typing. Init with dispatch init action.
The type saftey I wanted in the dispatcher wasn't doing much good as the
returned action could still differ from the action passed in so long as
the type was the same. Additionally, this typing prevented store
initialization because of type erasure. Removing the action typing
allows us to initialize the store properly and makes the dispatcher no
less safe. TODO: add documentation to Dispatcher to reflect this.